Add files via upload
We are given an mxn binary matrix of 1's (soldiers) and O's (civilians). The soldiers are positioned in front of the civilians. That is, all the 1's will appear to the left of all the O's in each row.
A row i is weaker than a row j if one of the following is true:
• The number of soldiers in row i is less than the number of soldiers in row j.
• Both rows have the same number of soldiers and i<j. Find the K weakest rows.
m=4, n=4, k=2
1000
1111
1000
1000